% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleTable-methods-exported.R
\name{SingleTable.modelFit}
\alias{SingleTable.modelFit}
\title{Exact posterior inference based on a single 2x2 table}
\usage{
SingleTable.modelFit(
  single_table_Obj,
  method = "exact",
  verbose = TRUE,
  control = list()
)
}
\arguments{
\item{sing_table_obj}{a list of components generated by \code{SingleTable.create()}.}
}
\description{
This function conducts exact posterior inference based on the list created by \code{SingleTable.create()}.
}
