% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleTable-methods-exported.R
\name{SingleTable.summary}
\alias{SingleTable.summary}
\title{Summarize the objects \code{singletable}}
\usage{
SingleTable.summary(
  single_table_Obj,
  alpha = 0.05,
  verbose = TRUE,
  digit = 3,
  control = list()
)
}
\arguments{
\item{alpha}{a numeric value specifying the significant level. Default value sets to 0.05.}

\item{verbose}{a logical value; if TRUE(default), the detailed summary messages will display.}

\item{digit}{an integer value specifying how many decimal places to keep.}

\item{control}{a list can be specified to control the fitting process.}

\item{sing_table_obj}{The object created by \code{SingleTable.create} and fitted by \code{SingleTable.modelFit}.}
}
\description{
Summarize a model of class \code{singletable} fitted by \code{SingleTable.modelFit}.
}
